# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Seletivo Seplag Gov-MT",
        "version": "0.1"
    },
    "paths": {
        "/api/album": {
            "get": {
                "tags": [
                    "Albuns"
                ],
                "summary": "Todos os Albuns",
                "description": "Lista todos os Albuns",
                "operationId": "7f7a3b4b6ddb653c4e41fac76cc46c75",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nº da página",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Albuns"
                ],
                "summary": "Cadastra novo Album",
                "description": "Registra um novo Album",
                "operationId": "7ef6c3ec7824931c32af2fe12f80514d",
                "parameters": [
                    {
                        "name": "artista_id",
                        "in": "query",
                        "description": "ID do artista",
                        "required": true
                    },
                    {
                        "name": "alb_titulo",
                        "in": "query",
                        "description": "Titulo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Album não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/album/{id}": {
            "get": {
                "tags": [
                    "Albuns"
                ],
                "summary": "Mostra um Album",
                "description": "Pesquisa por um album através do (id)",
                "operationId": "9b91d3bf023c51c24c0b2aad84a3f4d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Nº de identificação do album",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album Encontrado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Album não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Albuns"
                ],
                "summary": "Atualizar dados de um Album",
                "description": "Editar os dados de um album através do (id)",
                "operationId": "a389aa3f7a8e77e625a09a54e42ef1c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "artista_id",
                                    "alb_titulo"
                                ],
                                "properties": {
                                    "artista_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "alb_titulo": {
                                        "type": "string",
                                        "example": "Titulo album"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Album atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Album atualizado com sucesso"
                                        },
                                        "album": {
                                            "properties": {
                                                "artista_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "alb_titulo": {
                                                    "type": "string",
                                                    "example": "Nome Album"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Endereco não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Albuns"
                ],
                "summary": "Exclui um Album",
                "description": "Exclui um album através do (id)",
                "operationId": "09aee5d28508e14a14b9558226622b3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do album",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "album excluído com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Não foi possível excluir o album",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/album/{pesquisa}": {
            "get": {
                "tags": [
                    "Albuns"
                ],
                "summary": "Pesquisa Album pelo título",
                "description": "Pesquisa por um album através do titulo",
                "operationId": "2c3e3e10d1b34d338304489ce836539c",
                "parameters": [
                    {
                        "name": "pesquisa",
                        "in": "path",
                        "description": "Titulo de pesquisa do album",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "album"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Album Encontrado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Album não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}